Add can_be_renewed? to transient registration
Going forward we want to take the question of whether a registration can be renewed off the `CanChangeRegistrationStatus` concern and instead ask it of the transient registration instance itself.

So `can_be_renewed?` fits into the context of answering whether a renewal can be started, and continued, plus as it just deals current instance does not require any additional calls to the database.