Update forms ctrler to use trans reg's can_be_renewed?
Previously for every GET and POST to the site we would both query for the transient registration and then perform a subsequent query for the registration. Once we have the registration we could then call `may_renew?` on its metaData object.

With the changes we have made to Transient Registration and the addition of its `can_be_renewed?` method we can now make this call directly on the current instance of transient registration.