Update tests to fix cannot be renewed scenarios
Before the changes these tests relied on knowing that when we make a request, we first obtain the transient registration before then querying for the registration to carry out the can renew check.

With the changes we have made to `TransientRegistration` and the forms controller that second lookup no longer takes place.

Therefore we have had to update the test setup so that the transient registration itself is updated.