Block renewal rejection in initial conviction check state
https://eaflood.atlassian.net/browse/WC-483

When the conviction check is in its initial state, users should not be able to reject the renewal immediately, as anything with relevant convictions has to go through a more thorough process before rejection.

A few specs relied on immediately rejecting a conviction check, so these had to be updated.