Update completion service checks to be can complete
This is the second part of the refactoring. The checks we need to do when completing a renewal are different to that when starting or continuing a renewal application.

In the latter, its about whether then registration has expired before the declaration can be confirmed, or if the registration is still in the grace or renewal window.

When completing all we really care about is are there any outstanding conviction checks or payments, or has the registration been revoked since the application was submitted.

Hence the changes here update the RenewalCompletionService to reflect that. Also because `pending_payment?` is now part of the checks, we have had to do some refactoring in the tests and the factories to ensure that the orders and payments are done properly, and the balance is correct to allow the check to pass.