Fix failing test in travis ci
./spec/models/waste_carriers_engine/transient_registration_spec.rb:310 was failing in travis but not locally. The reason was that though we stub out the config value for the renewal window, code behind the scenes is also relying on there being a value for the grace window.

Locally there is because we have this set in our environment. On travis there isn't hence a different result.

Admittedly we should be deciding on a single route to go with these things (either stub all values or set all values via env vars before testing) however to keep things consistent for now this change updates the travis CI config to add the missing env var.