Create ready_to_complete? on trans registration
CodeClimate complained that the `RenewalCompletionService.can_be_completed?` had too high a cognitive complexity after our last changes.

And actually on review it was asking a number of questions of the transient registration instance it held, when really it should just be asking it whether it was ready to complete.

We still need the `can_be_completed?` method because of the additional check it makes on the status of the underlying registration. But now we resolve the code climate issue and put the checks where they really should be by creating a `TransientRegistration.ready_to_complete?` method.