Remove renew guard and move post updates to completion service
We've already moved the can renew checks to the transient registration object itself, so we need to delete the duplicate checks here.

Also, any changes to the registration should be managed by the RenewalCompletionService so we can see everything that happens in one place.

Finally as renewals will be driven through the renewal completion service, we remove the guard as the service will manage whether a registration can be renewed.