Move tests and delete duplicates
With all the changes to the CanChangeRegistrationStatus concern the tests based on whether calling metaData.renew would be blocked or not become redundant.

So this change either moves them to the transient registration or renewal completion service (depending on which is more appropriate), or deletes them if they duplicate something that already exists.