Refactor specs around renewal checks
One of the transient_registration specs was passing on Travis and some machines, but failing locally for me. After some investigation I discovered this was because the attempts to stub `Rails.configuration.grace_window` were failing, because that value was being called by the ExpiryCheckService, not the model being tested. Without the stub, the environment variable was always used, resulting in a failing test.

The solution to this issue is to make sure that the model spec is responsible for testing the model, and the service spec is responsible for testing the service. This PR refactors some specs and moves some others to achieve this.