Don't name form param when passing it to shared examples
We use shared examples heavily for GET and POST requests in our form pages. Previously we named each parameter we passed to the example, mostly for ease of reading - however, this triggers the 'useless assignment to variable' cop.

The form name is fairly self-explanatory, so we remove the assignment for that. However, other options will require further work.