Refactor tier decision for smart answers
This change is part of our ongoing housekeeping to enable Rubocop as part of our CI.

Rubocop flagged that the complexity of the method we used to check if a registration should be upper or lower tier based on the smart answers was high. So this PR attempts to solve this issue and make the code more readable.