Align rspec test setup with WEX
Whilst working on the WEX service where the code is based on WCR renewals, we have found issues with the initial Database cleaner setup, as well as how simplecov was configured.

We have also settled on a pattern where any test dependencies are required and configured in separate support files to keep the content of the `spec` and `rails` helpers to a minimum.

This change rectifies the issue with simplecov (the issue with Databasecleaner was already resolved), and implements the pattern, by aligning how rspec is setup to match the WEX service.