Handle whitespace when comparing Companies House numbers
https://eaflood.atlassian.net/browse/RUBY-240

When a registration is renewed, the user is asked to confirm their Companies House number. Then we compare this to the previous one. If they are different, we redirect the user to a page saying they need a new registration.

Sometimes registrations created in the old waste-carriers-frontend have trailing whitespace. The engine strips this whitespace automatically, but then this causes a discrepancy and the user cannot renew.

We have already submitted an RFC to fix the existing data - however to prevent this issue coming up again, we need to clean up trailing whitespace when comparing Companies House numbers.