Fix broken test and refactor expiry check service
Dependabot raised [PR 397](https://github.com/DEFRA/waste-carriers-renewals/pull/397) to update phonelib. However a test failed that broke the build.

```
1) WasteCarriersEngine::ExpiryCheckService#attributes when initialized with an upper tier registration :expiry_date matches the registration's
     Failure/Error: expect(subject.expiry_date).to eq(registration.expires_on)
       expected: Sun, 28 Mar 2021 09:49:03.121213803 +0000
            got: Sun, 28 Mar 2021 08:49:03.121213803 +0000
       (compared using ==)
       Diff:
       @@ -1,2 +1,2 @@
       -Sun, 28 Mar 2021 09:49:03 +0000
       +Sun, 28 Mar 2021 08:49:03 +0000
     # ./spec/services/waste_carriers_engine/expiry_check_service_spec.rb:38:in `block (4 levels) in <module:WasteCarriersEngine>'
```

After some local testing we were able to confirm that the issue was pre-existing and not because of phonelib.

It turns out that today (March 28 2019) is GMT, but in 3 years March 28 2021 will be in DST. This means what `ExpiryCheckService` stores as the `expiry_date` attribute is a corrected version and as such will not equal what is against the registration.

The very fact we know the `ExpiryCheckService` service corrects the date means the test is invalid; there must be times when `expiry_date` does not match `my_registration.expires_on` else what is the point of `ExpiryCheckService.corrected_expires_on`!?

So the ideal solution is that test determines whether the `expiry_date` should be back an hour, forward an hour or the same but we believe that would make it unnecessarily complex. It is sufficient to say that `expiry_date` is expected to be within  1 hour of the registration `expires_on` value.

Whilst we were looking at the code we also spotted some housekeeping improvements hence some refactoring changes in the `ExpiryCheckService` as well.