Fix error routes to contain a constraint for the /error path
Part of: https://eaflood.atlassian.net/browse/RUBY-261

The dummy app inside the WEX engine correctly handles 404 and 500 errors but the applications
that inherit from the engine always render a general template for every error.
This is because the wrong path is sent together as an ":id" and hence the information about the error template to use is lost in time and space.
Changing the URL constraint fixed this issue.
Hence the URL constraint is also going to change here for consistency and to prevent regression.