Fixes error handling redirect to internal error app
https://eaflood.atlassian.net/browse/RUBY-317

Added test coverage on content of the view and correct redirect from real error in requests error spec.
Fixed the return of the correct status code on the error controller.
The mounted error app initializer was already present.
Added spec helper to activate the correct error configs for testing purposes.
As agreed with Andrew, rendering the views is the main part. Missing paths, restyling and re-wording of the messages in the view will be tackled separately.