Replace built-in validator with DefraRuby::Validators one
One difference here is that the gem validator does not automatically pass if the transient_registration being validated doesn't need a `company_no`. So instead we just only call the validator when we know the value is required.

In the case of the registration number form, a user should not be on it unless the number is required, so we just skip the check.