Fix broken test
This refactor exposed a test which wasn't working as it should. We were setting a business_type to `sole_trader` in order to test the validation of our number of key people, but `sole_trader` was triggering a validation error itself as it should have been `soleTrader`. Once this was fixed, the test then failed and showed the validator wasn't working as expected. So this commit adds the necessary logic to the KeyPeopleValidator to create the expected behaviour and make the test pass.