Forms Refactoring - Add ability to reference single document in collection
From: https://eaflood.atlassian.net/browse/RUBY-693

In Wex, using Active Record, we were able to simplify a lot our logic around addresses forms by redefining the association to the address as `has_one` associations.
Currently, there is no way to do the same with MongId 5.2, hence this adds some metaprogramming code that will allow a mode l to define a single object that is persisted inside a location and to override it.