Fix bug in CanReferenceSingleDocument
For some reasons, when assigning an un-persisted object to a multiple association using `<<` and then saving the parent object, creates 2 transitions to the MongoDb API.
Since we are stuck with an old version of MongoId, the call is made using a deprecated attribute and an error is thrown:
```
  Mongo::Error::OperationFailure: Unknown modifier: $pushAll (9)
```
https://stackoverflow.com/questions/48607918/mongoerror-unknown-modifier-pushall-in-node-js/48621806#48621806

Before the change we were assigning the new object by re-assigning the entire hash in the collection. This will not trigger the `pushAll` command.

A new scenario have been added for documentation and to avoid regression,