Deal with special characters in search scope
https://eaflood.atlassian.net/browse/RUBY-696

This fixes a bug spotted in QA. Our existing search scope did not escape characters that were used in constructing regexes (eg `*`) which resulted in broken searches when the scope attempted to construct its own new regex.

This fix escapes the search term before using it in the regex, which should fix that problem.