Move setting up of ASSISTED_DIGITAL flag in AASM machine
Solve: https://eaflood.atlassian.net/browse/RUBY-718

We used the `RenewalCompletionService` to setup the metaData route of the transient_registration to the correct application-level configuration of `ASSISTED_DIGITAL` vs `DIGITAL`.
This was causing some registration to be incorrectly flagged as `ASSISTED_DIGITAL`, because the same service is run in the back-office application when a conviction check or payment check have been completed.
In order to make sure that we only flag the registration at the end of the forms workflow, we have moved the setting from the Service to the workflow AASM state machine.