Reset codeclimate to fix CI
Our codeclimate Ci integration seems to have stopped working for this project. With not really many options to diagnose why its stopped updating our PR status, we've gone nuclear!

So the changes to the code are minor (just updating the CI test key in the traviis config) behind the scenes we've deleted and re-added the project to codeclimate in an effort to resolve the issue.