Remove alias to `revoked` and add specific method
On the back-office details page, we use to show a warning message when conviction checks have been rejected.
In order to do so, we check that the status of the registration is `REVOKED`. This is not quite exact though, as we have specific ConvictionSignOff objects that deal with a rejection or approval.
In this PR I remove the alias method as that is not needed and instead create a more specific method in the transient registration which target the correct object status.