Remove order copy cards shared form and create concerns
In order to be able to reuse other forms in the order copy cards flow, such as the worldpay forms, this changes the strategy used to deal with the initialization of transient registrations in the base form and create a concern to deal with permission custom rules of the order copy cards form.
This will bring it more in line with what we have done in wex to share screens between the various transient objects (Edit, New, Renew).
However, this introduce a foundamental issue in the fact that transient registrations are not uniquely identifiable by reg_identifier any more. This means that this version of the code cannot be deployed until we have completed the refactoring which is going to replace reg_identifiers with unique tokens in routes and URL, as per WEX.