request/../conviction_details_forms_spec.rb passing
Main fix was ensuring the path was correct in the views.

Whilst working on this also spotted that some of our tests were not quite testing the right thing. Previouslyh we were passing in invalid params (first and last name are blank) along with an invalid reg identifier. With the help of @irisfaraway detective skills we could see that the reg identifier was always short circuting any validation checks and causing a redirect to 'Registration is invalid' page, hence the 302 code.

If you remove invalid reg identifiers from the equation (and in the new world tokens) what actually happens when invalid params like a blank first name and last name are submitted is that we return a 200 and simply `render :new`. So this change also removes any tests that were just checking for a 302 code.

N.B. It's likely we have other tests that make the same assumption being removed as I work through this.