requests/../renewal_complete_forms_spec passing
This required a tweak elsehwere than just the specs.

In the renewal completion service. Just needed to ensure we weren't trying to copy our new token field to the Registration instance.

The rest was spec changes. We'd been a bit to vigorous swapping `reg_identifier` for `token` so this fixed those.

The key change though is in tweaking the tests to match what now happens if a user is on the completion page and they hit refresh. Previously the code would have directed them to the renewal start page, where they then immediately got redirected to the unrenewable page.

Now though we're not passing a reg identifier, we're passing a token. This means the base form controlller validation tries to find the renewing transient registration, can't so ends up redirecting us to the invalid page. There's no obvious way we can get round this if we are going with using the token (plus I have a suspicion this is what we do in WEX as well for the same reasons).

So this change re-jigs the tests to try match the behaviour, and to try and reflect what is actually going on.