requests/../renewal_start_forms_spec passing
The changes here were mainly two-fold. The first was updating the controller to overrider the base form controller's `find_or_initialize_transient_registration()` with one that would consider the token as an actual token, or a registration identifier.

The second was then going back through the spec and ensuring we were referring to a token or a registration identifier, rather than using the token for everything.