Externalise permission checks logic into services
As part of building a new flow with shared forms with the existing renewal flow, I have stumped across the issue of how to deal with running the correct security checks when the same form is applied on different transient objects.
In order to solve the problem, we are now using a service to check the class of the transient object we are processing and based on the result, another service is run, specific to the transient object and flow, that will run the correct checks.