Remove hidden field from views
Based on how things were tracked previously, we had the reg identifier as a hidden field in each form. Now we rely on the token being the first part of the url so shouldn't need this hidden field.

In order to make the views cleaner this change removes those hidden fields.