Copy cards bank transfer form
From: https://eaflood.atlassian.net/browse/RUBY-767

This adds a new form and controller for the bank transfer form on the order copy cards journey.
It also includes a refactoring on the creation of the order payment, since the service was built to work on registration with existing finance details, it had to be modified so that it can work with transient objects.