Fix bug on creation of transient object
From: https://eaflood.atlassian.net/browse/RUBY-767

This fixes a bug which was creating a new transient object every time the start page of the flow was hit.
I am not sure this is actually a bug though, should we allow the user to start a new order every time? Might get them unblocked. But in that case we have to create a job to remove those transient registration as part of the release for this code, so that we can keep our database clean and avoid impacting the performance of the dashboard search.