Update BaseForm to work with Registrations
The back office has a few forms that inherit from BaseForm but supply a registration in place of a transient_registration. However, as registrations don't have tokens, this caused errors. This PR updates the BaseForm to only deal with tokens when the object in question is actually a TransientRegistration.

(We should arguably consider renaming `transient_registration` to `resource` or something like that, but I think that's a bigger job for another time!)