Use correct balance when displaying bank transfer info
From: https://eaflood.atlassian.net/browse/RUBY-767

The method I was using to show the user the correct amount to pay, is actually specific to renewals only.
I have added a TODO to the method so that we can move it in the specific object, and use the finance_details balance in this view.