Fix broken 'Finished' link on renewal end pages
https://eaflood.atlassian.net/browse/RUBY-859

We spotted that the 'Finished' button on the 'Renewal complete' and 'Renewal received' buttons was not behaving properly. For front office users, we now want to redirect straight to the front office dashboard; for back office users, we want to redirect to the new 'view details' page in the back office.

So we'll just provide a new default method in the engine and let the host apps override it as needed.