Set unique order key for non-worldpay payments
We have recently noticed that in the old system order keys on payments are randomly generated for all non-worldpay payments. We were under the impression that this was a minor details, hence in the new system we always use the last generated order key to set the payment's order code. The issue with that is that all the finance operations like refund and revert, rely on those codes to be unique as they are used to mark a payment as already refunded or reverted. Hence, currently, when a payment has been reverted, any other payment that has been created in the new system will not  be revertable as it has the same order code as the already reverted payment.