First attempt to fix databasecleaner update to 1.8
Currently we get the following error as soon as an attempt is made to run the specs.

```bash
An error occurred in a `before(:suite)` hook.
Failure/Error: DatabaseCleaner[:mongoid].strategy = :truncation

LoadError:
  cannot load such file -- database_cleaner/mongo/truncation_mixin
```

[Database Cleaner](https://github.com/DatabaseCleaner/database_cleaner) released 1.8.0 today (31 Jan 2020). It's been almost 2 years since their last release of 1.7.0 (19 April 2018).

From what I can ascertain a major part of the release is getting ready for all ORM adapters to be split off into their own gems ([Extract adapter gems](https://github.com/DatabaseCleaner/database_cleaner/pull/560)).

So the expectation now seems to be that rather than just define `gem "database_cleaner"` in your Gemfile, you need to be in explicit. In our case we use [Mongoid](https://github.com/mongodb/mongoid) so we need to

- update the Gemfile to `gem "database_cleaner-mongoid", "~> 1.8.0"`
- update `require "database_cleaner"` call where we configure Database Cleaner to be `require "database_cleaner/mongoid"`

So our first attempt to fix the error is to follow what we believe to be the new usage instructions.