Set up EditRegistration model
https://eaflood.atlassian.net/browse/RUBY-827

This PR sets up a new EditRegistration model to be used for editing. This will be a new type of TransientRegistration.

An EditRegistration should copy data from an existing Registration when it is created. It will also have its own workflow state machine. There should only be one EditRegistration in existence for a Registration at a time.