Store and display created_at for transient regs
https://eaflood.atlassian.net/browse/RUBY-829

The edit feature required us to store when an EditRegistration was created so we could display this information.

But I think it would be useful to store this for all our TransientRegistrations, as the absence of this info has caused problems in the past (for example, it was initially in the wireframe for the dashboard search results).

So this PR adds a little callback to set a value for `created_at` when we first persist a TransientRegistration.