Set up edit workflow
https://eaflood.atlassian.net/browse/RUBY-830

Editing is a less linear workflow than renewals or new registrations - we essentially want users to go from the overall edit page, to an existing form for an individual piece of data (eg contact phone) and then back to the overall edit page.

We should imitate the WEX implementation and behaviour.