Fix issue with missing expiry date
Lower tier registrations don't have an expiry date hence it blew up at this point. This resolves that matter and means the certificate now displays. However it also shows that the current view has been written solely with upper tier in mind.

This is understandable as that is all it was expected to cope until now.