Fix registered as section of certificate
The current registered as section was always assuming that the certificate was for an upper tier registration.

This change tweaks the locales and the presenter so it can handle both lower and upper tier registrations.

(Good news is this appears to be the last issue with displaying lower tier registrations!)