Add basic transitions for monodirectional states
By 'monodirectional' I mean you go from the edit overview, to this form, and then back to the edit. If you go 'back', you also go to the edit. So it's a straightforward A -> B -> A.

I pinched some shared examples and helper methods from WEX for this, which I'd like to start incorporating a bit more into our other transition tests, as they're pretty chonky!