Fix missing text values on edit page
https://eaflood.atlassian.net/browse/RUBY-829

I tested this out with some alternate business types and ran into problems. Some of these are down to dodgy data (charities shouldn't really have main people, for example) but we also needed to take into account some of the business types that are still generated by the frontend. Basically, we want to fail a little more gracefully!

Also swapped a `n/a` for `-` when a location is not provided, to be consistent with the registration details page.