Add routes for edit workflow
Basically these GET routes on the EditFormsController call the events on the state machine and then redirect to the correct form. So for example, the `edit_cbd_type` controller action will call the `edit_cbd_type` event in the state machine, which changes the state to `cbd_type_form`. The controller then redirects the user to the correct form.

This follows what we have done in WEX.