Run usual setup checks before doing edit transitions
We don't want to try to continue an edit if the user doesn't have the right permissions, or our other usual checks fail.

This also adds a bunch of other tests for circumstances when editing should not work.

To avoid an absolute apocalypse of request tests, I just tested the `cbd_type` ones as they are all so similar that I think this gives enough coverage.